requirements.txt
nicegui==2.5.0
tortoise-orm==0.21.7